moved fmt functions to own file
